Another migration fix
git-svn-id: https://ferma.svn.sf.net/svnroot/ferma/branches/double-coords-migration@317 d51c77f7-620d-0410-a109-9a2f07eaa567

+++ FermaNext/src/TrussUnit/TrussUnit.cpp
@@ -167,8 +167,8 @@ void TrussUnit::paintLoad ( const TrussLoad& load, const QPoint& tailPos,
     else if ( y > 0 )
         forceDirection.setY ( 1 );
 
-    QPoint headPos( tailPos.x() + 22 * x1 * forceDirection.x(), 
-                    tailPos.y() - 22 * y1 * forceDirection.y() );
+    QPoint headPos( int( tailPos.x() + 22 * x1 * forceDirection.x() ), 
+                    int( tailPos.y() - 22 * y1 * forceDirection.y() ) );
     
     solidRenderer solidRend ( baseRend );
     scanline_rasterizer   ras;
